% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/darnitools.R
\docType{package}
\name{darnitools-package}
\alias{darnitools}
\alias{darnitools-package}
\title{Tools for processing DARDAR-Nice data
TBC}
\description{
Just what the title says
}
\author{
\strong{Maintainer}: First Last \email{first.last@example.com} (YOUR-ORCID-ID)

}
