% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/grid_merge.R
\name{grid_merge}
\alias{grid_merge}
\title{Create gridded DARDAR-Nice data}
\usage{
grid_merge(dir_out)
}
\arguments{
\item{dir_out}{Directory used to output the gridded data}
}
\description{
Merge individual gridded orbits created by `darnitools::grid_orbit` into a single file
Grids created:
- global distributions of icnc_5um and icnc_100um (stored in dir_out/map)
- zonal distribtuions of icnc_5um and icnc_100um (stored in dir_out/zonal)
- Ni as function of T (stored in dir_out/ni_ta)
}
