% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/grid_merge.R
\name{grid_merge_map}
\alias{grid_merge_map}
\title{Create spatial distributions from DARDAR-Nice L2}
\usage{
grid_merge_map(dir)
}
\arguments{
\item{dir}{Directory used to output the gridded data}
}
\description{
Merge the files created by `grid_orbit_map`
}
