% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/grid_merge.R
\name{grid_merge_ni_ta}
\alias{grid_merge_ni_ta}
\title{Create Ni-Ta database}
\usage{
grid_merge_ni_ta(dir)
}
\arguments{
\item{dir}{Directory used to output the gridded data}
}
\description{
Merge the files created by `grid_orbit_ni_ta`
}
