% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/grid_merge_version.R
\name{grid_merge_version}
\alias{grid_merge_version}
\title{Compare two versions of the DARDAR-Nice L2 data using in grid}
\usage{
grid_merge_version(dir_gridded, v2)
}
\arguments{
\item{dir_gridded}{Directory where the gridded data is stored}

\item{v2}{Version number of the second version}
}
\description{
This function reads the gridded data created by
`grid_orbit_version` and merges all orbit files.
A rds file is created with the merged data.
}
