% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/grid_merge.R
\name{grid_merge_zonal}
\alias{grid_merge_zonal}
\title{Create zonal gridded means from DARDAR-Nice L2}
\usage{
grid_merge_zonal(dir)
}
\arguments{
\item{dir}{Directory used to output the gridded data}
}
\description{
Merge the files created by `grid_orbit_zonal`
}
