% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/grid_orbit.R
\name{grid_orbit}
\alias{grid_orbit}
\title{Grid DARDAR-Nice data from L2 files}
\usage{
grid_orbit(fn, dir_out, overwrite = FALSE)
}
\arguments{
\item{fn}{Name of the DARDAR-Nice L2 file}

\item{dir_out}{Directory used to output the gridded data}

\item{overwrite}{Overwrite existing deata. Default = FALSE}
}
\description{
This function takes a DARDAR-Nice L2 file and creates gridded data:
- global distributions of icnc_5um and icnc_100um (stored in dir_out/map/orbits)
- zonal distribtuions of icnc_5um and icnc_100um (stored in dir_out/zonal/orbits)
- Ni as function of T (stored in dir_out/ni_ta/orbits)
File function creates *.rds files that will need to be merged.
}
