% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/grid_orbit.R
\name{grid_orbit_map}
\alias{grid_orbit_map}
\title{Grid distributions of Ni}
\usage{
grid_orbit_map(
  df,
  fn_out,
  bins_lat = seq(-90, 90, by = 2),
  bins_lon = seq(-180, 180, by = 2),
  bins_ta = seq(-90, 0, by = 10)
)
}
\description{
Grid distributions of Ni
}
