% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/grid_orbit.R
\name{grid_orbit_ni_ta}
\alias{grid_orbit_ni_ta}
\title{Grid Ni as function of temperature}
\usage{
grid_orbit_ni_ta(
  df,
  fn_out,
  bins_ni = 10^seq(-3, 5, by = 0.05),
  bins_ta = seq(-90, 0, by = 1),
  season_region = FALSE
)
}
\description{
Grid Ni as function of temperature
}
