% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/grid_orbit_version.R
\name{grid_orbit_version}
\alias{grid_orbit_version}
\title{Grid DARDAR-Nice data from an orbit file between two versions}
\usage{
grid_orbit_version(
  fn_ref,
  v2,
  dir_out,
  bins_ta = seq(-90, 0, by = 10),
  overwrite = FALSE
)
}
\description{
Grid DARDAR-Nice data from an orbit file between two versions
}
