% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/grid_plot.R
\name{grid_plot_box_median}
\alias{grid_plot_box_median}
\title{Boxplot figure of Ni median for seasons and regions}
\usage{
grid_plot_box_median(fn, dir_out, suffix = NULL)
}
\arguments{
\item{fn}{File name of the DARDAR-Nice L3 data}

\item{dir_out}{Directory to save the figure}

\item{suffix}{Suffix to add to the file name}
}
\description{
Boxplot figure of Ni median for seasons and regions
}
