% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/grid_plot_version.R
\name{grid_plot_map_versions}
\alias{grid_plot_map_versions}
\title{Plot DARNI PRO L3 map differences between two versions}
\usage{
grid_plot_map_versions(fn_v1, fn_v2, dir_out, t_bins = c(-65, -45, -25))
}
\arguments{
\item{fn_v1}{File name of the first version gridded data}

\item{fn_v2}{File name of the second version gridded data}

\item{dir_out}{Directory to save the plot}

\item{t_bins}{Temperature bins to plot}
}
\description{
Plot DARNI PRO L3 map differences between two versions
}
