% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/grid_plot.R
\name{grid_plot_ni_ta}
\alias{grid_plot_ni_ta}
\title{Plotting DARDAR-Nice L3 Ni as function of temperature}
\usage{
grid_plot_ni_ta(fn, dir_out, suffix = NULL)
}
\arguments{
\item{fn}{File name of the DARDAR-Nice L3 data}

\item{dir_out}{Directory to save the plots}

\item{suffix}{Suffix to add to the file name}
}
\description{
Plotting DARDAR-Nice L3 Ni as function of temperature
}
