% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/grid_plot.R
\name{grid_plot_zonal}
\alias{grid_plot_zonal}
\title{Plotting DARDAR-Nice L3 zonal maps}
\usage{
grid_plot_zonal(fn, dir_out, ni_lim = c(0, 300), suffix = NULL)
}
\arguments{
\item{fn}{File name of the DARDAR-Nice L3 data}

\item{dir_out}{Directory to save the plots}

\item{ni_lim}{Limits of the color scale for the ice crystal number concentration (in #/L)}

\item{suffix}{Suffix to add to the file name}
}
\description{
Plotting DARDAR-Nice L3 zonal maps
}
