% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/grid_plot_version.R
\name{grid_plot_zonal_versions}
\alias{grid_plot_zonal_versions}
\title{Plot DARNI PRO L3 zonal differences between two versions}
\usage{
grid_plot_zonal_versions(fn_v1, fn_v2, dir_out)
}
\arguments{
\item{fn_v1}{File name of the first version gridded data}

\item{fn_v2}{File name of the second version gridded data}

\item{dir_out}{Directory to save the plot}
}
\description{
Plot DARNI PRO L3 zonal differences between two versions
}
