% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/l2.read.R
\name{l2.extract}
\alias{l2.extract}
\title{Extract necessary data from DARDAR-Nice L2 filenames (ta and icnc, mainly)
Data fame is adapted to run on some in situ functions}
\usage{
l2.extract(fn, dir.rds = NULL, only.ice = TRUE, keep.index = FALSE)
}
\arguments{
\item{fn}{name of DARDAR-Nice L2 file}

\item{dir.rds}{save output as an rds file if not NULL (default is NULL)}
}
\description{
Extract necessary data from DARDAR-Nice L2 filenames (ta and icnc, mainly)
Data fame is adapted to run on some in situ functions
}
