% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/l2_preview.R
\name{l2_preview}
\alias{l2_preview}
\title{DARDAR-Nice L2 preview}
\usage{
l2_preview(
  orbit.start,
  orbit.end,
  dir.data,
  dir.out = "~",
  return.data = FALSE
)
}
\arguments{
\item{orbit.start}{Start of the orbit to be processed (UTC). Format: POSIXct}

\item{orbit.end}{End of the orbit to be processed (UTC). Format: POSIXct}

\item{dir.data}{Directory where the data are stored}

\item{dir.out}{Directory where the output will be stored}

\item{return.data}{If TRUE, the function returns the data used to plot the preview.}
}
\description{
DARDAR-Nice L2 preview
}
