% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/l2_read.R
\name{l2_read}
\alias{l2_read}
\title{Read DARDAR-Nice L2 files}
\usage{
l2_read(
  fn,
  dir_rds = NULL,
  filter_ice = TRUE,
  filter_quality = TRUE,
  filter_cloudtop = FALSE,
  keep_index = FALSE
)
}
\arguments{
\item{fn}{Name of the DARDAR-Nice L2 file.}

\item{dir_rds}{Save output as an RDS file instead of returning the data. Default: NULL (return data).}

\item{filter_ice}{If TRUE, filter data to only include ice. Default: TRUE.}

\item{filter_quality}{Apply quality filters. Default: TRUE.}

\item{filter_cloudtop}{If TRUE, filter data to only include cloud top (first 180-m of uppermost ice layer). Default: FALSE.}

\item{keep_index}{If TRUE, keep index columns (idx_height and idx_time). Default: FALSE.}
}
\value{
A data frame containing the extracted and processed data or NULL if saved as RDS.
}
\description{
This function extracts DARDAR-Nice L2 files and returns a data frame or saves it as an RDS file.
The ice crystal concentration is converted from m-3 to L-1.
The temperature is converted from K to C.
The ice water content is converted from kg/m3 to mg/m3.
The effective radius is converted from m to um.
}
