% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/l2_read.R
\name{nc_read}
\alias{nc_read}
\title{Little tool to read netCDF files consistently}
\usage{
nc_read(nc, varname, idx = NULL)
}
\arguments{
\item{nc}{A netCDF file opened with ncdf4::nc_open()}

\item{varname}{The name of the variable to read}

\item{idx}{The specific index to associate. Default: NULL (output data as it is read)}
}
\value{
A numeric vector corresponding to the netCDF variable
}
\description{
This function is a wrapper around ncdf4::ncvar_get() to read netCDF files in a short and consistent way.
Each variables are transformed into 1D numeric vectors to be used in a data frame.
If the variable is 2D, the function can use the index of a specific dimension to fist extrapolate 1D variables into 2D grid before transforming them into 1D vectors.
The index can also be used to find specific values after the initial data was filtered.
}
