% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/set_panel_size.R
\name{set_panel_size}
\alias{set_panel_size}
\title{Useful utility for setting the panel size, margin, width, and height of a ggplot object.
It saves the plot as an image if a file path is specified.}
\usage{
set_panel_size(
  p = NULL,
  g = ggplot::ggplotGrob(p),
  file = NULL,
  margin = grid::unit(1, "mm"),
  width = grid::unit(4, "cm"),
  height = grid::unit(4, "cm")
)
}
\arguments{
\item{p}{A ggplot object.}

\item{g}{A ggplot grob. Default: ggplotGrob(p).}

\item{file}{File path to save the plot as an image. Default: NULL (do not save).}

\item{margin}{Margin around the plot. Default: unit(1, "mm").}

\item{width}{Width of the panel. Default: unit(4, "cm").}

\item{height}{Height of the panel. Default: unit(4, "cm").}
}
\value{
A ggplot object with specified panel size, margin, width, and height.
}
\description{
Useful utility for setting the panel size, margin, width, and height of a ggplot object.
It saves the plot as an image if a file path is specified.
}
