% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/darnitools.R
\docType{package}
\name{darnitools-package}
\alias{darnitools}
\alias{darnitools-package}
\title{Tools for processing DARDAR-Nice data
TBC}
\description{
Basically what the title says.
}
\author{
\strong{Maintainer}: Odran Sourdeval \email{odran.sourdeval@univ-lille.fr}

}
