% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/grid_merge.R
\name{grid_merge}
\alias{grid_merge}
\title{Create gridded DARDAR-Nice data}
\usage{
grid_merge(dir_out)
}
\arguments{
\item{dir}{Directory used to output the gridded data}
}
\description{
Merge the grids from individual orbital files
}
