% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/l2_grid.R
\name{l2_grid}
\alias{l2_grid}
\title{Overall function to grid the DARDAR-Nice data}
\usage{
l2_grid(
  dir_darni,
  dir_gridded,
  years_only = NULL,
  version_2 = NULL,
  overwrite = FALSE,
  nthreads = 1
)
}
\arguments{
\item{dir_darni}{Directory containing DARDAR-Nice L2 data.}

\item{dir_gridded}{Directory that will contain the gridded product.}

\item{years_only}{Years to process. Default: NULL (all years).}

\item{version_2}{Version 2 of the DARDAR-Nice data to compare with. Default: NULL (no comparison).}

\item{overwrite}{Overwrite the data if it exists. Default: FALSE.}

\item{nthreads}{Number of parallel threads. Default: nthreads = 1.}
}
\description{
This function is a wrapper around the functions: \code{\link{grid_orbit}} and \code{\link{grid_merge}}.
If \code{version_2} is not NULL, it will also call \code{\link{grid_orbit_version}} and \code{\link{grid_merge_version}}.
}
