% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/l2_grid.R
\name{l2_grid}
\alias{l2_grid}
\title{Grid the DARDAR-Nice data}
\usage{
l2_grid(dir_darni, dir_gridded, overwrite = FALSE, nthreads = 1)
}
\arguments{
\item{dir_darni}{Directory containing DARDAR-Nice L2 data.}

\item{dir_gridded}{Directory that will contain the gridded product.}

\item{overwrite}{Overwrite the data if it exists. Default: FALSE.}

\item{nthreads}{Number of parallel threads. Default: nthreads = 1.}
}
\description{
Grid the DARDAR-Nice data
}
